% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered.R
\name{centering}
\alias{centering}
\title{Calculates the center of the normal dist}
\usage{
centering(prob)
}
\arguments{
\item{prob}{The desired probability of success}
}
\value{
Where to center the normal distribution
}
\description{
Using the logistic function, takes a desired proportion / probability of
success and returns the mean of the distribution that will likely generate
probability close to desired proportions.
}
